// web/src/app/core/api/admin-generated/models/content-generation-status-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AuthoringRevisionStatus } from '../models/authoring-revision-status';
import { ContentGenerationJobStatus } from '../models/content-generation-job-status';
export interface ContentGenerationStatusResponse {
  attemptCount?: (number | string);
  createdAt?: string;
  errorCode?: (string | null);
  errorMessage?: (string | null);
  finishedAt?: (string | null);
  jobId?: string;
  jobStatus?: ContentGenerationJobStatus;
  revisionId?: string;
  revisionStatus?: AuthoringRevisionStatus;
  startedAt?: (string | null);
}