// web/src/app/core/api/admin-generated/models/generated-suite-review-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { GeneratedCaseReviewResponse } from '../models/generated-case-review-response';
import { SuiteQualityPolicy } from '../models/suite-quality-policy';
export interface GeneratedSuiteReviewResponse {
  casePreview?: Array<GeneratedCaseReviewResponse>;
  casesByGroup?: {
[key: string]: (number | string);
};
  isCasePreviewTruncated?: boolean;
  killedCaseCountByWrongSolution?: {
[key: string]: (number | string);
};
  qualityPolicy?: SuiteQualityPolicy;
  revisionId?: string;
  suiteSha256?: string;
  survivingWrongSolutions?: Array<string>;
  testCaseCount?: (number | string);
  toolchain?: string;
  wrongSolutionCount?: (number | string);
}