// web/src/app/core/api/admin-generated/models/problem-draft-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AuthoringRevisionStatus } from '../models/authoring-revision-status';
import { DifficultyLevel } from '../models/difficulty-level';
import { ProblemAuthoringDefinition } from '../models/problem-authoring-definition';
import { ProblemSampleRequest } from '../models/problem-sample-request';
export interface ProblemDraftResponse {
  constraintsMarkdown?: string;
  definition?: ProblemAuthoringDefinition;
  difficulty?: DifficultyLevel;
  memoryLimitKb?: (number | string);
  problemId?: (number | string);
  revisionId?: string;
  revisionNumber?: (number | string);
  samples?: Array<ProblemSampleRequest>;
  slug?: string;
  statementMarkdown?: string;
  status?: AuthoringRevisionStatus;
  timeLimitMs?: (number | string);
  title?: string;
  updatedAt?: string;
}